test: isolate add_blocks schema helpers
- move recursive schema assertions into a dedicated test helper module

- remove the unused schema type helper from docx-specific test utilities

- share the public string schema used by add_blocks list and table payloads
