+++ tests/_docx_test_helpers.py
@@ -72,33 +72,6 @@ def _find_paragraph(doc, text: str):
     return next(paragraph for paragraph in doc.paragraphs if paragraph.text == text)
 
 
-def _schema_contains_key(schema: object, key: str) -> bool:
-    if isinstance(schema, dict):
-        return key in schema or any(
-            _schema_contains_key(value, key) for value in schema.values()
-        )
-    if isinstance(schema, list):
-        return any(_schema_contains_key(value, key) for value in schema)
-    return False
-
-
-def _schema_type_allows(schema: dict, expected_type: str) -> bool:
-    schema_type = schema.get("type")
-    if isinstance(schema_type, list):
-        return expected_type in schema_type
-    return schema_type == expected_type
-
-
-def _schema_contains_type_list(schema: object) -> bool:
-    if isinstance(schema, dict):
-        if isinstance(schema.get("type"), list):
-            return True
-        return any(_schema_contains_type_list(value) for value in schema.values())
-    if isinstance(schema, list):
-        return any(_schema_contains_type_list(value) for value in schema)
-    return False
-
-
 def _paragraph_field_codes(paragraph) -> list[str]:
     from docx.oxml.ns import qn
